Refactor buildingLayoutView
