--- src/screens/buildingDetail/BuildingLayoutViewModel.ts
@@ -1,12 +1,0 @@
-import { ImageSourcePropType } from 'react-native'
-
-export interface BuildingLayoutViewModel {
-  buildingTypeName: string
-  buildingTypeIcon: ImageSourcePropType
-  layout: HouseLayoutViewModel
-}
-
-export interface HouseLayoutViewModel {
-  actionTitle: string
-  actionSubtitle: string
-}
